Move Throttler class into correct package
